Add general improvements

+++ include/comms.h
@@ -138,7 +138,7 @@ int request_finalize(ucp_worker_h ucp_worker, send_req_t *request, send_req_t *c
 size_t send_req(ucp_worker_h ucp_worker, ucp_ep_h ep, ucp_address_t *addr, size_t addr_len, char * request);
 size_t send_data(ucp_worker_h ucp_worker, ucp_ep_h ep, const void *msg, size_t msg_len, uint64_t from);
 size_t get_recv_data_length(ucp_worker_h ucp_worker,  uint64_t dest);
-int32_t recv_data(ucp_worker_h ucp_worker, ucp_ep_h ep, void *msg, size_t msg_length,  uint64_t dest, int async);
+size_t recv_data(ucp_worker_h ucp_worker, ucp_ep_h ep, void *msg, size_t msg_length,  uint64_t dest, int async);
 size_t recv_req(ucp_worker_h ucp_worker, ucp_ep_h ep, char *msg);
 ucs_status_t request_wait(ucp_worker_h ucp_worker, void *request, send_req_t *ctx);
 void stream_recv_cb(void *request, ucs_status_t status, size_t length, void *user_data);
